Fixed assorted compiler warnings.

+++ jumpdrive/RPNExpression.cpp
@@ -17,7 +17,7 @@ RPNExpression::RPNExpression(const list<Token*> &tokens, const string &expressio
 ValueSet RPNExpression::calculate(const ValueSet values) {
 	if (m_variables.size() == 0 && values.size() > 0)
 		throw new IllegalArgumentException("there are no variables to set values");
-    else if (values.size() != m_variables.size())
+    else if ((unsigned int)values.size() != m_variables.size())
 		throw new IllegalArgumentException("The are an unequal number of variables and arguments");
 
 	if (m_variables.size() > 0 && !values.empty()) {
